Refactor ExpressionMonotonicityVisitor
