+++ common/planners/src/lib.rs
@@ -32,6 +32,7 @@ mod plan_expression_column;
 mod plan_expression_common;
 mod plan_expression_function;
 mod plan_expression_literal;
+mod plan_expression_monotonicity;
 mod plan_expression_rewriter;
 mod plan_expression_sort;
 mod plan_expression_validator;
@@ -116,6 +117,7 @@ pub use plan_expression_function::neg;
 pub use plan_expression_function::not;
 pub use plan_expression_function::sum;
 pub use plan_expression_literal::lit;
+pub use plan_expression_monotonicity::ExpressionMonotonicityVisitor;
 pub use plan_expression_rewriter::ExpressionRewriter;
 pub use plan_expression_sort::sort;
 pub use plan_expression_validator::validate_expression;
