+++ query/src/optimizers/mod.rs
@@ -19,7 +19,6 @@ mod optimizer_expression_transform;
 mod optimizer_scatters;
 mod optimizer_statistics_exact;
 mod optimizer_top_n_push_down;
-mod utils;
 
 pub use optimizer::Optimizer;
 pub use optimizer::Optimizers;
@@ -28,4 +27,3 @@ pub use optimizer_expression_transform::ExprTransformOptimizer;
 pub use optimizer_scatters::ScattersOptimizer;
 pub use optimizer_statistics_exact::StatisticsExactOptimizer;
 pub use optimizer_top_n_push_down::TopNPushDownOptimizer;
-pub use utils::MonotonicityCheckVisitor;
