+++ query/src/optimizers/optimizer_top_n_push_down.rs
@@ -22,7 +22,6 @@ use common_exception::Result;
 use common_planners::*;
 use common_tracing::tracing;
 
-use super::MonotonicityCheckVisitor;
 use crate::optimizers::Optimizer;
 use crate::sessions::QueryContext;
 
@@ -172,7 +171,7 @@ impl TopNPushDownImpl {
                     Some((l, r)) => (l.clone(), r.clone()),
                 };
 
-                match MonotonicityCheckVisitor::extract_sort_column(
+                match ExpressionMonotonicityVisitor::extract_sort_column(
                     schema.clone(),
                     expr,
                     left,
